+++ src/include/duckdb/function/scalar/string_functions.hpp
@@ -10,6 +10,7 @@
 
 #include "duckdb/function/scalar_function.hpp"
 #include "duckdb/function/function_set.hpp"
+#include "utf8proc.hpp"
 
 namespace re2 {
 class RE2;
@@ -45,6 +46,22 @@ struct ConcatWSFun {
 
 struct LengthFun {
 	static void RegisterFunction(BuiltinFunctions &set);
+	template <class TA, class TR> static inline TR Length(TA input) {
+		auto input_data = input.GetData();
+		auto input_length = input.GetSize();
+		for (idx_t i = 0; i < input_length; i++) {
+			if (input_data[i] & 0x80) {
+				int64_t length = 0;
+				// non-ascii character: use grapheme iterator on remainder of string
+				utf8proc_grapheme_callback(input_data, input_length, [&](size_t start, size_t end) {
+					length++;
+					return true;
+				});
+				return length;
+			}
+		}
+		return input_length;
+	}
 };
 
 struct LikeFun {
@@ -69,6 +86,9 @@ struct RegexpFun {
 
 struct SubstringFun {
 	static void RegisterFunction(BuiltinFunctions &set);
+	static string_t substring_ascii_only(Vector &result, const char *input_data, int offset, int length);
+	static string_t substring_scalar_function(Vector &result, string_t input, int offset, int length,
+	                                          unique_ptr<char[]> &output, idx_t &current_len);
 };
 
 struct PrintfFun {
